+++ src/Models/DomainModel.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace {{ namespace }};
+namespace Lunarstorm\LaravelDDD\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Lunarstorm\LaravelDDD\Factories\DomainFactory;
 use Illuminate\Database\Eloquent\Model;
+use Lunarstorm\LaravelDDD\Factories\DomainFactory;
 
-abstract class {{ class }} extends Model
+abstract class DomainModel extends Model
 {
     use HasFactory;