+++ src/Support/DomainResolver.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Lunarstorm\LaravelDDD\Support;
+
+class DomainResolver
+{
+    public static function fromModelClass(string $modelClass)
+    {
+    }
+}