+++ src/ValueObjects/DomainObject.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Lunarstorm\LaravelDDD\ValueObjects;
+
+class DomainObject
+{
+    public function __construct(
+        public readonly string $name,
+        public readonly string $namespace,
+        public readonly string $fqn,
+        public readonly string $path,
+    ) {
+    }
+}