+++ tests/Datasets/Domains.php
@@ -6,3 +6,9 @@ dataset('domainPaths', [
     ['Custom/PathTo/Domain', 'Domain'],
     ['Custom/PathTo/Domains', 'Domains'],
 ]);
+
+dataset('domainSubdomain', [
+    // Domain, Subdomain
+    ['Customer', 'Reporting'],
+    ['Customer', null],
+]);